Added standard parse HTML function. FIxes #94
