+++ src/templates/blocks/blocks/text.js
@@ -1,8 +1,6 @@
 import React from 'react'
 import styled from '@emotion/styled'
-import LinkInspect from 'components/link-inspect'
-import parse from 'html-react-parser'
-import domToReact from 'html-react-parser/lib/dom-to-react'
+import parseHtml from '../parse-html'
 
 const Paragraph = styled('p')`
   ${props =>
@@ -13,23 +11,11 @@ const Paragraph = styled('p')`
       : ``};
 `
 
-const parserOptions = {
-  replace: ({ type, name, attribs, children }) => {
-    if (type === 'tag' && name === 'a') {
-      return (
-        <LinkInspect to={attribs.href}>
-          {domToReact(children, parserOptions)}
-        </LinkInspect>
-      )
-    }
-  },
-}
-
 const BlockText = ({ lead, text }) => {
   if (!text) {
     return null
   }
-  return <Paragraph lead={lead}>{parse(text, parserOptions)}</Paragraph>
+  return <Paragraph lead={lead}>{parseHtml(text)}</Paragraph>
 }
 
 export default BlockText
